Dev/vftable

+++ src/modules/tag/tag.service.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { TagRepository } from './tag.repository';
 import { RedisService } from 'src/common/redis/redis.service';
 import { NoTagException } from './exception/no-tag.exception';
+import { Interval } from '@nestjs/schedule';
 
 const TAG_CACHE_KEY = 'tagRedisCache@@@@@@';
 
@@ -20,10 +21,12 @@ export class TagService {
     await this.tagRepository.insertTag(accountIdx, tag, songIdx);
   }
 
+  @Interval(60000 * 60)
   async cacheTagAll() {
     const data = await this.tagRepository.selectTagAll();
     await this.redisService.set(TAG_CACHE_KEY, JSON.stringify(data));
   }
+
   async findTagAllByCache() {
     const data = await this.redisService.get(TAG_CACHE_KEY);
     if (!data) {
@@ -33,6 +36,11 @@ export class TagService {
   }
 
   async deleteTag(tagIdx: number) {
+    const tagCheck = await this.tagRepository.selectTagByIdx(tagIdx);
+    if (!tagCheck) {
+      throw new NoTagException();
+    }
     await this.tagRepository.deleteTagByIdx(tagIdx);
+    await this.cacheTagAll();
   }
 }
